% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_uri_keystage.R
\name{format_uri_keystage}
\alias{format_uri_keystage}
\title{Format keystage query for dfe-api}
\usage{
format_uri_keystage(stage = NULL, filter = NULL, order = NULL,
  skip = NULL, take = NULL)
}
\arguments{
\item{\code{stage}}{Keystage of interest: takes either `ks2`, `ks4`,
or `ks5`. deafults to `NULL`.}

\item{\code{skip}}{Takes an integer; these many records will be
skipped over and no returns.}

\item{\code{take}}{How many records to return. Although the default in
this function is `NULL`, the serverside default is 50.}
}
\value{
A properly formatted uri string for querying the API, for example:
`http://dfe-api.azurewebsites.net/api/keystage/search/ks4?`.
}
\description{
\code{format_uri_keystage} formats a queri uri string
ready for submission to dfe-api.
}
\details{
Takes a properly formatted uri string as input, and makes a
request of dfe-api, parses the returning JSON, and returns a list.
}
\examples{

\dontrun{

library(dfeAPI)
library(dplyr)

 # Try basic functionality

format_uri_keystage(
   stage = "ks4",
   take = 10
) \%>\%
   get_data

}

}

