% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get data from dfe-api}
\usage{
get_data(uri_string = NULL)
}
\arguments{
\item{\code{uri_string}}{A properly formatted uri string with which to
query the API. This would principally come from \code{format_uri()}
and its relatives.}
}
\value{
A list containing the response, typically containing: ErrorMessage,
NumberOfResults, Results, Taken, Skipped, Facets, QueryLat, QueryLong,
Disambiguate, DisambiguationSuggestions
}
\description{
\code{get_data} The main function behind \code{dfeAPI}.
}
\details{
Takes a properly formatted uri string as input, and makes a
request of dfe-api, parses the returning JSON, and returns a list.
}
\examples{

\dontrun{

library(dfeAPI)
library(dplyr)

response <- "ks4" \%>\%
   format_uri \%>\%
   get_data

 # Convert to dplyr's data_frame (as data are very wide!)

 foo <- response$results \%>\%
   as_data_frame

   }

}

